implement test_constructors for Timestamp (#31583)
